Reorganize `group/*`
