+++ src/l2hmc/group/su3/tensorflow/utils.py
